View model, note

+++ app/src/main/java/com/online/languages/study/lang/NoteActivity.java
@@ -1,192 +1,124 @@
 package com.online.languages.study.lang;
 
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.PreferenceManager;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.widget.Toolbar;
-
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.ImageView;
-import android.widget.TextView;
-
+import androidx.appcompat.app.AlertDialog;
+import androidx.databinding.DataBindingUtil;
+import androidx.lifecycle.ViewModelProvider;
 import com.online.languages.study.lang.adapters.InfoDialog;
 import com.online.languages.study.lang.adapters.OpenActivity;
 import com.online.languages.study.lang.adapters.RoundedCornersTransformation;
-import com.online.languages.study.lang.adapters.ThemeAdapter;
 import com.online.languages.study.lang.data.DataManager;
 import com.online.languages.study.lang.data.NoteData;
+import com.online.languages.study.lang.databinding.ActivityNoteBinding;
+import com.online.languages.study.lang.view_models.NoteViewModel;
+import com.online.languages.study.lang.view_models.NoteViewModelFactory;
 import com.squareup.picasso.Picasso;
 
 import static com.online.languages.study.lang.Constants.ACTION_UPDATE;
 import static com.online.languages.study.lang.Constants.EXTRA_NOTE_ACTION;
 import static com.online.languages.study.lang.Constants.EXTRA_NOTE_ID;
 
-public class NoteActivity extends BaseActivity {
-
-
-    ThemeAdapter themeAdapter;
-    SharedPreferences appSettings;
-    public String themeTitle;
+public class NoteActivity extends ThemedActivity {
 
     OpenActivity openActivity;
-    TextView title, titleNopic, content;
-    ImageView noteIcon;
-    View titleWrap;
 
     String source = "";
 
-    String noteId;
-    DataManager dataManager;
     NoteData note;
 
-    String [] note_pics = new String[] {};
-    String folder = "";
-
+    ActivityNoteBinding binding;
+    NoteViewModel noteViewModel;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-
-        appSettings = PreferenceManager.getDefaultSharedPreferences(this);
-        themeTitle= appSettings.getString("theme", Constants.SET_THEME_DEFAULT);
-
-        themeAdapter = new ThemeAdapter(this, themeTitle, false);
-        themeAdapter.getTheme();
-
         openActivity = new OpenActivity(this);
         openActivity.setOrientation();
 
-        noteId = getIntent().getStringExtra(EXTRA_NOTE_ID);
-
         source = getIntent().getStringExtra("source");
-        if (source  == null) {
+        if (source == null) {
             source = "";
         }
 
-        note_pics = getResources().getStringArray(R.array.note_pics_list);
-        folder = getString(R.string.notes_pics_folder);
+        String noteId = getIntent().getStringExtra(EXTRA_NOTE_ID);
 
+        noteViewModel = new ViewModelProvider(this, new NoteViewModelFactory(this.getApplication(), noteId)).get(NoteViewModel.class);
 
-        setContentView(R.layout.activity_note);
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_note);
+        binding.setLifecycleOwner(this);
+        binding.setNoteViewModel(noteViewModel);
 
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        setSupportActionBar(binding.toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         setTitle(R.string.note_title_txt);
-
-        title = findViewById(R.id.noteTitle);
-        titleNopic = findViewById(R.id.noteTitleNopic);
-        titleWrap = findViewById(R.id.titleWrap);
-        content = findViewById(R.id.noteContent);
-        noteIcon = findViewById(R.id.noteIcon);
-
-
-        dataManager = new DataManager(this);
 
         setNote();
 
     }
 
 
-
     private void setNote() {
 
-        note = dataManager.dbHelper.getNote(noteId);
+        // note = dataManager.dbHelper.getNote(noteId);
 
-        String titleTxt = note.title;
-        String titleContent = note.content;
+        note = noteViewModel.loadNote();
 
-        title.setText(titleTxt);
-        titleNopic.setText(titleTxt);
-        content.setText(titleContent);
-        noteIcon.setVisibility(View.VISIBLE);
+        manageTitleAndIcon();
 
+        String imagePath = noteViewModel.imagePath();
 
-        String noteImage = checkImage(note.image);
-
-        if (emptyImage(noteImage)) {
-            titleNopic.setVisibility(View.VISIBLE);
-            noteIcon.setVisibility(View.GONE);
-            title.setVisibility(View.GONE);
-        } else {
-            titleNopic.setVisibility(View.GONE);
-            noteIcon.setVisibility(View.VISIBLE);
-            title.setVisibility(View.VISIBLE);
-        }
-
-        if (note.title.equals("") && emptyImage(noteImage)) {
-            titleWrap.setVisibility(View.GONE);
-        } else {
-            titleWrap.setVisibility(View.VISIBLE);
-        }
-
-        Picasso.with( this )
-                .load("file:///android_asset/pics/" +folder + noteImage )
+        Picasso.with(this)
+                .load(imagePath)
                 .fit()
                 .centerCrop()
-                .transform(new RoundedCornersTransformation(20,0))
-                .into(noteIcon);
+                .transform(new RoundedCornersTransformation(20, 0))
+                .into(binding.noteIcon);
 
     }
 
 
-    private String checkImage(String picName) {
+    private void manageTitleAndIcon() {
 
-        boolean found = false;
-        if (picName == null) picName = "";
+        boolean emptyImage = noteViewModel.emptyImage();
 
-        String img = picName;
-
-        for (String pic: note_pics) {
-
-            if (picName.equals(pic)) {
-                found = true;
-                break;
-            }
+        if (emptyImage) {
+            binding.noteTitleNopic.setVisibility(View.VISIBLE);
+            binding.noteIcon.setVisibility(View.GONE);
+            binding.noteTitle.setVisibility(View.GONE);
+        } else {
+            binding.noteTitleNopic.setVisibility(View.GONE);
+            binding.noteIcon.setVisibility(View.VISIBLE);
+            binding.noteTitle.setVisibility(View.VISIBLE);
         }
 
-        if (!found) img = "none";
-
-        return img;
-    }
-
-
-
-
-    private boolean emptyImage(String picName) {
-
-        boolean noImage = false;
-
-        if (picName.equals("none") || picName.equals("empty.png") || picName.equals("")) {
-            noImage = true;
+        if (note.title.equals("") && emptyImage) {
+            binding.titleWrap.setVisibility(View.GONE);
+        } else {
+            binding.titleWrap.setVisibility(View.VISIBLE);
         }
-
-        return noImage;
     }
 
 
     private void editNote() {
         Intent i = new Intent(this, NoteEditActivity.class);
-        i.putExtra(EXTRA_NOTE_ID, note.id );
-        i.putExtra(EXTRA_NOTE_ACTION, ACTION_UPDATE );
-
+        i.putExtra(EXTRA_NOTE_ID, note.id);
+        i.putExtra(EXTRA_NOTE_ACTION, ACTION_UPDATE);
         startActivityForResult(i, 20);
     }
 
+
     private void deleteNote() {
+
+        DataManager dataManager = new DataManager(this);
         dataManager.dbHelper.deleteNote(note);
         finish();
     }
-
-
-
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
@@ -198,32 +130,32 @@ public class NoteActivity extends BaseActivity {
     }
 
 
-
     @Override
     public void onBackPressed() {
         super.onBackPressed();
         openActivity.pageBackTransition();
     }
 
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
-        switch(id) {
-            case android.R.id.home:
-                finish();
-                openActivity.pageBackTransition();
-                return true;
-            case R.id.edit_note:
-                editNote();
-                return true;
-            case R.id.delete_note:
-                deleteConfirmDialog();
-                return true;
-            case R.id.info_item:
-                showInfo();
-                return true;
 
+        if (id == android.R.id.home) {
+            finish();
+            openActivity.pageBackTransition();
+            return true;
+        } else if (id == R.id.edit_note) {
+            editNote();
+            return true;
+        } else if (id == R.id.delete_note) {
+            deleteConfirmDialog();
+            return true;
+        } else if (id == R.id.info_item) {
+            showInfo();
+            return true;
         }
+
         return super.onOptionsItemSelected(item);
     }
 
@@ -231,10 +163,9 @@ public class NoteActivity extends BaseActivity {
     @Override
     public void finish() {
 
-
-        if (source.equals("search"))  {
+        if (source.equals("search")) {
             Intent returnIntent = new Intent();
-            setResult(RESULT_OK,returnIntent);
+            setResult(RESULT_OK, returnIntent);
             returnIntent.putExtra("position", getIntent().getIntExtra("position", 0));
         }
 
@@ -243,16 +174,9 @@ public class NoteActivity extends BaseActivity {
     }
 
 
-
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu_note, menu);
-
-        if (source.equals("search")) {
-           // menu.findItem(R.id.edit_note).setVisible(false);
-            //menu.findItem(R.id.delete_note).setVisible(false);
-        }
-
         return true;
     }
 
@@ -260,9 +184,7 @@ public class NoteActivity extends BaseActivity {
     private void showInfo() {
         InfoDialog infoDialog = new InfoDialog(this);
         infoDialog.simpleDialog(getString(R.string.info_notes_title), getString(R.string.info_notes_text));
-
     }
-
 
     public void deleteConfirmDialog() {
 
@@ -270,26 +192,12 @@ public class NoteActivity extends BaseActivity {
         builder.setTitle(R.string.confirmation_txt);
         builder.setMessage(R.string.delete_note_confirm);
         builder.setCancelable(true);
-
-        builder.setPositiveButton(R.string.continue_txt, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                deleteNote();
-
-            }
-        });
-
-        builder.setNegativeButton(R.string.cancel_txt, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-
-            }
+        builder.setPositiveButton(R.string.continue_txt, (dialog, which) -> deleteNote());
+        builder.setNegativeButton(R.string.cancel_txt, (dialog, which) -> {
         });
 
         builder.show();
 
     }
-
-
 
 }
